requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-secret-manager
requests
